+++ circuits/CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update `poseidon-merkle` to v0.9.0-rc.0
 - Update `dusk-poseidon` to v0.42.0-rc.0
 - Update `jubjub-schnorr` to v0.7.0-rc.0
-- Update `jubjub-elgamal` to v0.5.0-rc.0
+- Update `jubjub-elgamal` to v0.5.0-rc.1
+- Use `encrypt_unchecked` elgamal gadget [#280]
 
 ## [0.6.0] - 2025-02-07
 
@@ -136,6 +137,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update `poseidon-merkle` to v0.6 [#179]
 
 <!-- ISSUES -->
+[#280]: https://github.com/dusk-network/phoenix/issues/280
 [#255]: https://github.com/dusk-network/phoenix/issues/255
 [#235]: https://github.com/dusk-network/phoenix/issues/235
 [#232]: https://github.com/dusk-network/phoenix/issues/232